divide into different components, more organized
