+++ client/src/users/styles/Profile.css
@@ -172,7 +172,6 @@
   color: black;
   text-align: center;
   line-height: 45px;
-  /* background-color: aqua; */
   text-decoration: none;
   transition: 0.25s;
 }
